ECS implementation

+++ src/helpers/Constants.h
@@ -1,5 +1,5 @@
 #pragma once
 
 constexpr uint16_t FPS = 60;
-constexpr uint16_t TIME_PER_FRAME = 1000 / FPS;
+constexpr int TIME_PER_FRAME = 1000 / FPS;
 constexpr uint16_t MAX_COMPONENTS = 32;
